requirements.txt
openai==1.79.0
opencv_python==4.11.0.86
Pillow==11.2.1
pyktok==0.0.31
python-dotenv==1.1.0
python-telegram-bot==22.1
Requests==2.32.3
vosk==0.3.45
